Materialized views for aggregated data
This change helps speed up the read of ancillary data (i.e. build,
classification) by pre-aggregating data periodically.

Additionally model.py was moved from shared directory to telemetryui
directory since no other component depends on this module.
